+++ Core/EventArgs/PortfolioUpdateEventArgs.cs
@@ -3,6 +3,6 @@
     public class PortfolioUpdateEventArgs : System.EventArgs
     {
         public object Type { get; internal set; }
-        public double Value { get; internal set; }
+        public decimal Value { get; internal set; }
     }
 }
