+++ Core/PriceActionDetectors/LevelDetector.cs
@@ -8,10 +8,10 @@ namespace Core.PriceActionDetectors
 {
     public class LevelDetector
     {
-        public IEnumerable<double> Levels;
+        public IEnumerable<decimal> Levels;
         private SortedDictionary<int, int> map = new();
 
-        public void Process(double price)
+        public void Process(decimal price)
         {
             int p = (int)price;
             if (!map.ContainsKey(p))
@@ -24,7 +24,7 @@ namespace Core.PriceActionDetectors
                 map[p]++;
             }
 
-            Levels = map.OrderBy(x => x.Value).Take(3).Select(x => (double)x.Key);
+            Levels = map.OrderBy(x => x.Value).Take(3).Select(x => (decimal)x.Key);
         }
     }
 }
